+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Cinecred Changelog
 ### Additions
 
 - Added higher quality profiles for H.264 and H.265 delivery.
+- Added Theora delivery.
 - Completed render jobs now show their total runtime.
 
 ### Fixes